simulator/requirements.txt
paho-mqtt==2.1.0
loguru==0.7.2
